+++ src/system/kernel/debugger/LocalIO.h
@@ -41,8 +41,8 @@ public:
    * from either the top or bottom of the screen, respectively. Can be used to 
    * create status lines that aren't destroyed by screen scrolling.
    */
-  void setCliUpperLimit(int nlines);
-  void setCliLowerLimit(int nlines);
+  void setCliUpperLimit(size_t nlines);
+  void setCliLowerLimit(size_t nlines);
 
   /**
    * Enables or disables the command line interface, allowing full access to the display.
@@ -60,23 +60,23 @@ public:
    * horizontal or vertical direction. Note that if the CLI is enabled,
    * anything drawn across the CLI area can be wiped without warning.
    */
-  void drawHorizontalLine(char c, int row, int colStart, int colEnd, DebuggerIO::Colour foreColour, DebuggerIO::Colour backColour);
-  void drawVerticalLine(char c, int col, int rowStart, int rowEnd, DebuggerIO::Colour foreColour, DebuggerIO::Colour backColour);
+  void drawHorizontalLine(char c, size_t row, size_t colStart, size_t colEnd, DebuggerIO::Colour foreColour, DebuggerIO::Colour backColour);
+  void drawVerticalLine(char c, size_t col, size_t rowStart, size_t rowEnd, DebuggerIO::Colour foreColour, DebuggerIO::Colour backColour);
 
   /**
    * Draws a string of text at the given location in the given colour.
    * note that wrapping is not performed, the string will be clipped.
    */
-  void drawString(const char *str, int row, int col, DebuggerIO::Colour foreColour, DebuggerIO::Colour backColour);
+  void drawString(const char *str, size_t row, size_t col, DebuggerIO::Colour foreColour, DebuggerIO::Colour backColour);
 
   /**
    * Returns the width and height respectively of the console.
    */
-  int getWidth()
+  size_t getWidth()
   {
     return CONSOLE_WIDTH;
   }
-  int getHeight()
+  size_t getHeight()
   {
     return CONSOLE_HEIGHT;
   }
@@ -110,23 +110,23 @@ protected:
   /**
    * Framebuffer.
    */
-  unsigned short m_pFramebuffer[CONSOLE_WIDTH*CONSOLE_HEIGHT];
+  uint16_t m_pFramebuffer[CONSOLE_WIDTH*CONSOLE_HEIGHT];
 
   /**
    * Framebuffer for the screen before we took control.
    */
-  unsigned short m_pOldFramebuffer[CONSOLE_WIDTH*CONSOLE_HEIGHT];
+  uint16_t m_pOldFramebuffer[CONSOLE_WIDTH*CONSOLE_HEIGHT];
 
   /**
    * Current upper and lower CLI limits.
    */
-  int m_UpperCliLimit; /// How many lines from the top of the screen the top of our CLI area is.
-  int m_LowerCliLimit; /// How many lines from the bottom of the screen the bottom of our CLI area is.
+  size_t m_UpperCliLimit; /// How many lines from the top of the screen the top of our CLI area is.
+  size_t m_LowerCliLimit; /// How many lines from the bottom of the screen the bottom of our CLI area is.
 
   /**
    * Current cursor position.
    */
-  int m_CursorX, m_CursorY;
+  size_t m_CursorX, m_CursorY;
 
   /**
    * Keyboard state.
